partial build 1.3, without payment gateway

+++ app/src/main/java/com/mobile/cloudkitchen/data/model/Rating.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 
 data class Rating(
 
-    @SerializedName("avgRating") var avgRating: Int? = null,
-    @SerializedName("ratings") var ratings: ArrayList<Int> = arrayListOf(),
+    @SerializedName("avgRating") var avgRating: Double? = null,
+    @SerializedName("ratings") var ratings: ArrayList<Double> = arrayListOf(),
     @SerializedName("noOfRatings") var noOfRatings: Int? = null
 
 )
